[VMware Updates] Added Light Italic, Semibold weights; Added trademark and registration symbols; Fixed Capital 'S' issue
_Describe the change or feature in detail. Try to answer the following questions._

* What are the changes? Tell us why for each one, if possible.
Added Semibold and Light Italic weight.

Addition of both the Trademark symbol (™) and the Registration symbol (®) to all weights. These characters are needed for representing properties or entities that are trademarks or registered trademarks of an individual or corporation.

Adjusted the sizing of the capital "S" to be more consistent with the sizing of the rest of the capital letterforms.

* Why should it go in Metropolis?
The Trademark and Registration symbols are often required by the legal departments of corporations to clarify legal ownership of intellectual property.
The adjusted size of the capital "S" makes upper case treatments more consistent across all letterforms.

_Provide some specific details about how this change would improve your application and use case._

If referencing products that are the registered intellectual property of a person or corporation, it is necessary to add the Trademark or Registration symbol (as applicable) on first mention. This can also be done by writing out "Product X is a registered trademark of Person or Corporation Y", however using the symbols saves space and is a best-practices preferred treatment.

When using an all upper case type style, the uniformity of the capital S creates a cleaner final typographic treatment.